introducing es6 rules and modules to the awesome books app
